logging frontend works
