+++ backend/src/server.js
@@ -15,6 +15,7 @@ const userRouter = require('./serverRouter/userRouter');
 const groupRouter = require('./serverRouter/groupRouter');
 const workgroupsRouter = require('./serverRouter/workgroups');
 const storyRouter = require('./serverRouter/storyRouter');
+const logging = require('./logging')
 require('./database/DbServices');
 
 const app = express();
@@ -66,66 +67,13 @@ app
 	}));
 
 
-const winston = require('winston')
-function getLogger(){
-	//Winston config
-	const myformat = winston.format.combine(
-		winston.format.colorize(),
-		winston.format.timestamp(),
-		winston.format.align(),
-		winston.format.printf(info => `${info.timestamp} ${info.level}: ${info.message}`)
-	);
-	const logConfiguration = {
-		transports: [
-			new winston.transports.Console({
-				level: 'debug',
-				format: myformat
-			}),
-			new winston.transports.File({
-				level: 'warn',
-				filename: './logs/backend_warn.log',
-				format: myformat
-			}),
-			new winston.transports.File({
-				level: 'debug',
-				filename: './logs/backend_debug.log',
-				format: myformat
-			})
-		]
-	};
-	return winston.createLogger(logConfiguration);
-}
-const logger = getLogger();
 app
 	.use(passport.initialize())
 	.use(passport.session())
 	.use(bodyParser.json({ limit: '100kb' }))
 	.use(bodyParser.urlencoded({ limit: '100kb', extended: true }))
 	.use((_, __, next) => {
-		if (_.url.endsWith('log')) 
-		{next()}
-		else{
-			console.log('Time:', Date.now());
-			let current_datetime = new Date();
-			let formatted_date =
-				current_datetime.getFullYear() +
-				"-" +
-				(current_datetime.getMonth() + 1) +
-				"-" +
-				current_datetime.getDate() +
-				" " +
-				current_datetime.getHours() +
-				":" +
-				current_datetime.getMinutes() +
-				":" +
-				current_datetime.getSeconds();
-			let method = _.method;
-			let url = _.url;
-			let status = __.statusCode;
-			let log = `[${formatted_date}] ${method}:${url} ${status}`;
-			logger.debug(log)
-			next();
-		}
+		logging.httpLog(_, __, next)
 	})
 	.use('/api/script', scriptRouter)
 	.use('/api/run', runReportRouter)
